Do not require signature for trusted ID tokens
OPs may return unsigned tokens in response to a token request made by Sync Gateway (auth code or refresh flows).  These tokens will be invalid for client authentication (via bearer token), but when received from the token endpoint can be treated as valid identity confirmation.  In this scenario, create and return a valid session to the client.  The unsigned token is also returned to the client, as it may still be useful to client applications (extracting claim information).

The test provider is enhanced to simulate this scenario, by setting the unsigned_id_token property in the SG config:
```
        "unsupported": {
          "oidc_test_provider": {
            "enabled":true,
            "unsigned_id_token":true
          }
        }
```