Issue #1925: Obscure HTTP basic auth credentials in webhook URLs (#1932)
* Issue #1925: Obscure HTTP basic auth credentials in webhook URLs

A webhook event's URL may have HTTP basic auth credentials embedded directly into it (e.g. http://username:password@example.com/events), which was, but is no longer, exposed as plain text when the webhook is included in a log message (e.g. when the webhook is first registered or a webhook event is processed). Instead the username-and-password component of a webhook's URL is now obscured with asterisks (e.g. http://****:****@example.com/events) when it is converted to a string.

While it may have been simplest to remove the webhook URL from such log messages altogether, I thought it best to leave it in but in a sanitized format so that it's possible to distinguish between different event types when a bucket/database has more than one webhook enabled. Otherwise webhook events are effectively indistinguishable from each other without knowing the URL.

* Issue #1925: Obscure HTTP basic auth creds in additional instances of webhook URLs

There are a couple more cases where the full webhook URL is written to Sync Gateway's log. They are now sanitized to obscure HTTP basic auth credentials too.