Fix wait handling when changes runs ahead of stable sequence
Queries against the channel index can return values later than the stable sequence; we catch scenario in index_changes and don't send to clients.  However, this had the potential to leave the changes feed stuck in a wait loop until a subsequent write happened for that channel.  To prevent this, bypass wait loop processing when we see a post-stable sequence.