Changes unit test fix, user handling fix
A newly changed unit test (to add wait handling) created a deadlock issue.  Added a timeout to the changes request for that test to avoid the deadlock, and realized that the actual cause of the deadlock was a recently added new exit point to the changes feed that wasn't sending an error to the feed runner on termination.  Corrected that handling as well.