Fixes #2266 loading persisted walrus buckets CBL ios test failure (#2281)
If the Walrus PR is merged via "rebase and merge", then this PR should already be pointing to the exact commit that will land on the walrus master branch.  Otherwise if "squash" or "normal merge" is used, we'll have to rework this PR to point to the correct commit in the walrus repo.