Refactor backfill processing to handle edge cases (#2283)
Fix for https://github.com/couchbaselabs/sync-gateway-accel/issues/76.

The previous backfill processing had gaps when the changes feed was interrupted by limit.  Refactored to better follow the model used for non-vector sequences - once a channel is in backfill, we send everything pre-since for that channel, without interference from other non-backfill channels.

The key change is to defer calculation of the backfill results until the changes feed reaches the backfill trigger (instead of calculating at the start of the changes loop iteration).  This allows the backfill to be based on the since value at trigger time, and ignore anything later than the since value.