Prevent backfill for triggers later than stable sequence
Channel backfill is skipped when the triggering sequence is later than the stable sequence, to ensure that the triggering sequence is included with the backfill (avoids duplicate backfill).

Also updates polling to remove activeReaders check - since one-shot changes are using the cached results, need to refresh the stable sequence even when no continuous changes are in progress.