Concurrency fixes in partition storage
On a cas failure during update of the active block, we were attempting to reapply the changes to the current version of the block.  This didn't handle the case where the active block had changed.

Fix is to reload the block list and re-initialize the active block before attempting to reapply the changes.