Standardize dense storage list persistence (#2317)
DenseStorageLists are stored as JSON, but we were manually converting to []byte then sending that as {}interface to WriteCas.  Technically this should work - gocb's transcoding checks the format of the incoming {}interface, but switching to the simpler JSON object handling in case something in the transcoding was resulting in the mangled document format.

This has the additional advantage of using the existing retry handling on Get, which might also help if this is the result of a transient error.