Notify when sequences are allocated and not used
When Sync Gateway fails to update a document after an unsuccessful WriteUpdate (CAS write loop), it ends up with an unused allocated sequence.  This causes Sync Gateway nodes that are reading the TAP/DCP feed and trying to generate a set of ordered sequences to block waiting for that sequence (eventually triggering the skipped sequence handling).

Successful WriteUpdate retries were already handling this case using the unusedSequences property on the successfully updated doc.

On failed WriteUpdate with unused sequences, we now persist a transient document named _sync:unusedSeq:[seqnumber].  When Sync Gateway nodes working the TAP/DCP feed see a document of this type, the sequence gets buffered as an unused sequence.