PR Feedback: found an issue with error comparison
I just noticed an issue while looking into this.

Currently we're calling:

if err != nil && err != gocb.ErrKeyNotFound {
but in my testing I found that this comparison doesn't work, since gocb.ErrKeyNotFound is actually a pointer and it doesn't seem to provide any guarantees that it will be returning a pointer to the same object. I didn't actually look at the code, but from my testing that comparison approach wasn't reliable, which is why I switched to the string comparison approach.

I will push up a fix that uses the string comparison.