Add purge support
Currently purge has to be done as a two-step operation (delete body, then delete xattr) until server delivers the capability to do both in one operation.  To ensure this doesn't break anything in the meantime, import is cancelled for delete mutations if we can't find an existing _sync xattr.

Adds check in change_cache.go to skip import processing when the echo of the purge comes over the DCP feed.

Includes a fix in GetWithXattr to properly return KeyNotFound when both xattr and doc don't exist. (+3 squashed commits)
Squashed commits:
[7c6b45a] more working
[87118ce] Working
[c2b4f25] Fix incr handling for empty key (#2540)

On an empty key, incr was returning error instead of '0' after singleOps refactoring