Trigger doc import on SG write
If a document has been updated by a non-SG author but has not yet been imported, trigger import on an attempt to update the doc via Sync Gateway. Fixes #2554.

Refactored the import processing to return the imported doc, so that on-demand imports can proceed without having to requery the bucket for the document.  Normal cas check will handle the case where the document is updated by someone else after import but before subsequent processing.

This cleans up the on-demand import during a GET (no longer requires loop), and avoids an unnecessary kv operation when on-demand import is triggered by a PUT.

Also fixed handling in change_cache.go when node is running in xattr mode but not importing.  Previously non-SG writes would attempt to get cached, then be rejected based on the stale sequence value.  Instead, we shouldn't even try to cache in the first place.