Avoid Sync Gateway panic when unmarshaling null document body
null is valid JSON, and previously returned a nil map (not an empty map) when unmarshalled as Body.  This caused subsequent attempted writes to the map to panic.

Sync Gateway doesn't support null documents - modified the unmarshalling to return an error in this scenario, which will result in the doc being ignored by Import/Shadowing.