Compact obsolete tombstones from accel's channel index
When a tombstone is purged by Couchbase Server, no DCP notification is issued.  To remove obsolete tombstones from the index bucket, accel needs to do the work to track tombstones and remove them.

To identify tombstones that should be compacted, and the set of channels those tombstones are associated with, accel now maintains a doc-based tombstone index in the index bucket.  Each time accel indexes a tombstone in the channel index, it now also writes a tombstone index entry to this index, specifying the tombstone time, the vbucket and sequence, and the set of channels the tombstone was indexed to.  See tombstone.go for implementation details.

Periodically a scheduled process queries that index (for vbuckets assigned to the accel node) to identify any tombstones that are older than the server's metadata purge interval.  These tombstones are then purged from the channel index and the new tombstone index.  This process runs every 24h by default, but can be customized using the new channel_index.tombstone_compact_freq property in the config.