Fix AccessWarning unit test
Test was already broken - the attempt to add the principal failed due to missing password.  There may have been some walrus context bleed allowing it to pass - it passed when run along with the full suite of AccessWarning tests, but failed when run alone (on master).

Fixed to close server contexts in previous tests, corrected principal creation, and added retry loop to account for stale=ok.