Enable tombstone indexing for Sync Gateway design docs
View Engine now requires an additional property (index_xattr_on_deleted_docs) to be set when creating a design doc, if any views in that design doc need to index tombstones.  This property is intended to be internal, and so isn't supported via the SDK (gocb).

Extracted the relevant code into bucket_gocb to allow Sync Gateway to write a design doc while setting the new property.

Includes an additional fix to move ViewPrincipals into the housekeeping design doc, as it's no longer a core processing view.