Mobile opt-in using import filter
Allows definition of an javascript import filter function in the Sync Gateway config, which is used to determine whether an SDK write should be available to mobile (imported).  The filter function takes the document body as an input parameter, and is expected to return a boolean to indicate whether the document should be imported.

The filter is defined in the Sync Gateway config under a new 'import_filter' property on the database configuration.

 Additional details on the opt-in use cases here: https://docs.google.com/document/d/1XxLIBsjuj_UxTTJs4Iu7C7uZdos8ZEzeckrVc17y3sw

Added a new import.go to help organize the import-related functionality.