Add expiry() built-in to sync function
Calling expiry(value) from within the sync function will set the expiry value for a document, in the same way as if the _exp property were set in the document.

In addition to adding the new function to sync runner, required modification to the Update and WriteUpdate bucket methods to support passing the (potentially) modified expiry back along with the document body.

Also refactored sg-bucket and it's implementations to standardize expiry as uint32 - matches gocb and the underlying Couchbase binary protocol, and cleans up inconsistent usage/casting in SG (originally due to the go-couchbase API using int).