Sporadic Test Fix - TestLowSequenceHandling
Modifies the behaviour of TestLowSequenceHandling and TestLowSequenceHandlingAcrossChannels to only validate that the expected sequences are received, and ignore race-related variability in total number of changes received or low sequence values.

Moved the previous TestLowSequenceHandling (with the duplicate check intact) to 'DisableTestLowSequenceHandlingNoDuplicates', and updated #3056 to indicate that it should be re-enabled when that issue is fixed.