Adjust maxRecentSequence handling
Previously it pruned when number of recent sequences persisted to the doc exceeded kMaxRecentSequences, which meant that a given doc could store kMaxRecentSequences+1.  Switched to prune at >= kMaxRecentSequences, which is more intuitive behaviour.

Unit test that was previously asserting incorrectly (based on unintuitive behaviour) could fail intermittently (when it happened to land on kMaxRecentSequences+1) - should now pass consistently.