Inclusive_end should be true if not specified (#3077)
* Inclusive_end should be true if not specified

Default value of inclusive_end in Couchbase Server, if not specified, is true.

* Fix active-only loop in getChangesInChannelFromView

The active-only query loop wasn't terminating when endSeq (representing the data already in the cache) was reached.  This resulted in a query being issued with startKey greater than endKey.  This causes Couchbase Server to return an error when inclusive_end is true.

* Add unit test for #3078