Don't store incoming _revisions property in rev cache
The only internal properties that should be included in the rev cache version of the body are _id and _rev.  Modified the API on RevisionCache.Put to make this explicit - callers now provide a docid, revid, and body as is stored in the bucket.  _id and _rev are inserted into the body as it's stored in the rev cache.  Ensures any other internal properties (like _revisions) that aren't persisted to the bucket also aren't stored in the rev cache.

Fixes #3107