Reload active user context on change to their role's channel set (#3116)
* Reload active user context on change to their role's channel set

When a user writes a document that triggers a change to the set of channels on a role they have been granted, reload the user context.  This ensures that the user context is updated to include the changed channels for subsequent writes in the same bulk_docs request.

* Standardize usage of "role:" prefix

Also cleans up some test formatting based on PR feedback.