Updated go-blip, and added BLIP fatal-error handler that logs to SG
Picked up bug-fixes in go-blip, plus the ability to set a FatalError
handler.

Added a FatalError handler that logs to "HTTP". This will show things
like go-blip panics, or fatal protocol errors, in the SG log.

Also took the error logging out of the conn.Close() call in the defer
block -- I found that the only errors that it ever logged were
"closing an already-closed socket", if go-blip had already closed
the connection, so the log call was just noise.

Fixes #3192