Ensure changes consistency during cache eviction (#4146)
Adds handling to ensure that the same channel cache instance is used for the late feed and normal feed during a given changes iteration, and strengthens the check to guarantee that the active late handler is associated with the current cache instance.  In case of mismatch, triggers the normal feed to roll back to the stable sequence.