GetChanges logging fix (#4254)
In channelCache.GetChanges was failing to output info logging in the case where resultFromCache was empty but non-nil.

As previously written, the 'nothing cached' log line was unreachable.  This was resulting in unexpected info logging output
when c.GetCachedChanges returned a non-nil empty slice.  In particular, a changes request that queried two channels -  one
that was empty and anotherl that had no results for the requested range - would log "0 changes valid from' for the former, and log nothing
for the latter, which gave the impression the second channel wasn't being queried.