TLS client certificate support
If enabled via `unsupported.client_cert_auth` config, TLS client certs
will be used for authentication. A client cert overrides normal auth
of all requests sent on its socket. A configured pattern is used to
look up a username from the info in the cert, and if the database has
a user with that name, the request is authenticated as that user.
Other authentication (HTTP Basic, session cookie, etc.) is ignored.

Configuration changes:

If `unsupported.client_cert_auth` is present, it must be an object
containing the following keys:
* `state`, whose value must be "disabled", "enabled", or "mandatory".
* `ca_cert`, whose value is the path to a PEM or DER
file containing the CA certificate used to validate client certs.
* `prefixes`, whose value is an array of objects; for details see:
https://docs.google.com/document/d/1sC_He6DZdiZBw63jIvOdwqD_2uGAkELzSA1BqxGuNCA

CBG-570