Add changeWaiter to TestReloadUser (#4345)
With the userWaiter changes to how blip users are reloaded, there is a potential race between access grants occurring *outside* of a given replication, and the user being reloaded.  The user now won't be reloaded until the invalidation of the user's channels arrives over DCP.  (note that grants that originate from the same replication will still refresh the user immediately).  Updates the test to cover this race.