Reduce map lookup/iterate in processEntry (#4383)
Use slice in addToCache, and only use map where needed for union-type operations.

Requires an additional check in AddToCache for the case where a document is explicitly added to the star channel.  Also removed info level logging for “negative latency”, as this is common in multi-node scenarios with small clock skew.