CBG-623 Avoid duplicate marshalling during write/import during rev id generation (#4392)
* Avoid duplicate marshalling during write/import during rev id generation

Leverage existing body for rev id generation (import) and retain marshalled body for subsequent write when appropriate.

* Test fixes

Previous TestImportNullDoc was making an invalid call to importDoc (body and existingDoc.body both provided, didn't match).  Removed the invalid case, and added a new test for the scenario that the previous test was attempting to cover (and added some additional comments on the expected parameter usage for importDoc).