CBG-870: TestOpenIDConnectAuth occasionally failing on Build Jenkins and under -race
When creating the refresh response the expiry value was calculated as "time.Now().Add(5 * time.Minute).UTC().Second()" that always returns the second offset within the minute specified by time, in the range [0, 59]. When Now + 5 minutes qualifies a 0 value for second offset (eg.2020-05-19 14:44:00), expiry is set in the response as 0 and when unmarshalling the response at the caller side, the zero value expiry is not getting set in the OIDCTokenResponse object. Eventually, the test ends up in expiry mismatch assertion failure since the previous expiry value of OIDCTokenResponse received during auth response was not cleared.

- Removed the assertion to validate the refresh token during refresh flow since refresh token is not included in the response and asserting the refresh token received during the auth response doesn't really make sense.
- Set the value of token expiry in both auth and refresh responses to make it sync with the expiry value set in JWT claims.