CBG-930 Delta sync flag for sg-replicate2
If delta sync is disabled for an sg-replicate replication, override the calculated sgCanUseDeltas value for the replication's blip sync context.

Note that this implies that pull replications from a remote that supports delta sync will still not use deltas unless deltas are enabled on the local node.  This matches CBL behaviour today (we don't accept deltas from CBL if deltas aren't enabled for the SG database).