Throttle concurrent in-flight changes requests per connection to 2
Couchbase Lite already limits the number of in-flight changes batches
to prevent Sync Gateway from preparing the revs for batches too far ahead of
time and having to keep them in memory before we're able to send them.

This commit enforces that limit on the server side for each client, to
prevent non-CBL clients from issuing too many changes requests before
the revs for those batches have had chance to be sent.

We keep this number >1 in order to keep the list of revs supplied for
clients to allow for a continous stream of revisions to be available and
prevent "batchiness" in the sync.