CBG-1127 Don't treat imported SDK resurrections as conflicts
When a tombstoned document is resurrected via the SDK, the metadata history is lost and the document is resurrected as a new revision on a disconnected branch.  When replicated via sg-replicate, this revision shouldn't be treated as being in conflict with a tombstoned document (no active leaves).

Concurrent updates and deletes to the same revision continue to be treated as conflicts, as this doesn't meet the 'disconnected branch' criteria.