CBG-1910 Avoid retrieval of temporary revision backups during ancestor lookup
When attempting to retrieve an ancestor for use as oldDoc in the sync function, we shouldn't check for temporary revision body backups.  These aren't intended for use by the sync function (as they would provide different results depending on whether they have expired), and the check adds unnecessary performance overhead at write time when the ancestor is unavailable or not the immediate parent.