CBG-2010 Avoid attachment fetch work for known attachments independent of revpos
For blip replication, we previously relied on revpos to identify that an attachment was already associated with the Sync Gateway version of the document.  However, in cases where revpos isn't specified correctly, SG can still avoid this work by checking existing attachments.

If an incoming attachment matches name and digest with an attachment on the active server revision of the document, we no longer perform prove/get work, even when revpos is greater than the common ancestor.

This also addresses an issue where we were checking against v2 attachment storage was always being used when revpos > minRevPos, under the assumption that this must be a new attachment to SG and should be stored in the new format. For clients using revpos incorrectly, this resulted in attachments being migrated from v1 to v2 storage unintentionally.