CBG-2030 Fix limit handling for _user endpoint
The ordering of ORDER BY, LIMIT and the startKey predicate were causing a failure when the result set was larger than the limit and the pagination handling kicked in.  Reorganized the query to include the startKey predicate in all cases, since QueryPrincipals is the only current use case and an empty startkey is handled properly.  Added test coverage for the low limit case, and moved the existing TestUserAPI test to a new file for better functional organization.