CBG-2027 Enhancement to GET /db/_user/
Adds the ability to include email and disabled values when retrieving all users via GET /db/_user/ by setting name_only=false.  Also supports specifying a limit on that REST API when name_only=false is being used.

This is to support this functionality in 3.0.1 for Capella, so views support is not required (and not included, as it would require an update to view definitions).  For 3.1.0 we plan to support this functionality with new indexes on the system collection.