CBG-2050 Add option for read-only anonymous access
Adds an internal (unsupported) option to make anonymous access (via GUEST user) read-only.  Includes low-level enforcement immediately prior to sync function execution, but also performs an earlier check for 2.x and REST API calls.  REST API calls are identifying read-only calls based on the admin WriteAppData permission - this overloads that functionality somewhat, and would be a candidate for a wider refactoring of privileges for the REST API in future.

Also adds a missing userBlipHandler for MessageGetRev and MessagePutRev.