CBG-2121: Forward port CBG-2055: Do not evaluate revpos during storeAttachments (#5557)
* CBG-2055 Do not evaluate revpos during storeAttachments

Revpos shouldn't be used as a criteria for attachment persistence for incoming writes. In particular during storeAttachments, a variety of circumstances can result in changes to the incoming revpos, and we don't want to fail to persist or remove attachments based on revpos.

In addition to removing the revpos check in storeAttachments, adds some additional handling in blip_handler for different flavors of revpos being sent by CBL to ensure it's being set to the incoming revID when the attachment doesn't exist.

* Update rest/blip_api_test.go

Co-authored-by: Ben Brooks <ben.brooks@couchbase.com>

* Ensure version is accurate for new attachments with shared digests

* Add test for legacy attachment name change

* Additional test case for legacy attachment w/out name change

Co-authored-by: Ben Brooks <ben.brooks@couchbase.com>