User query: param substitution in config channel names
In the user query config, the channel names in the "channels" property can now
contain patterns of the form `$param` or `$(param)` that will be replaced by
the value of the corresponding parameter at runtime. This allows the channel
access required for a query to vary based on the query parameters.

- As usual, the `$` has to be doubled as `$$` in the config file to avoid
  environment-variable substitution.
- If needed, a literal `$` in a channel name can be achieved by doubling it,
  which in the config file would look like `$$$$`.
- `param` must be alphanumeric, unless parenthesized.
- The value of the parameter must be a string or integer, otherwise the
  query fails with a 400 status.
- As of this commit, you can access the user name as `$user`, but not the other
  sub-properties of `user`.