Narrow the lock in fetchAndLoadConfigs to avoid high response latency
When running the periodic fetchAndLoadConfigs for discovering updated
persistent database configs, we acquire the ServerContext lock before
making the network calls to discover buckets and fetch config documents.

This can take a while (~3 seconds) due to having to list all buckets,
and then for each fetch a config document, and we don't need to be
holding the ServerContext lock to do this work.

Moving the lock down after we've done the network hops means we're only
holding the ServerContext lock when we're applying the configs we got
from the buckets, which is a relatively short operation.