GraphQL: Implemented "__typename" resolvers for interface values
Without this feature, queries that used values of an interface type would fail
because they couldn't figure out what concrete object type the value was.
For example:
```
	node(id:"route_10003") {
		id
    		... on Route {
          equipment
        }
    }
```
The `Query.node(id)` resolver returns type Node, an interface. This query needs
to detect whether the actual value is a Route or not, but until this commit the
GraphQL engine had no way of knowing. After this commit, you can add a resolver
for `Node. __typename ` to the config:
```
"Node": { "__typename": `
    switch (value.type) {
        case "route": return "Route";
...
```