Refactored function/query code into new 'functions' package
The new file db/functions.go now defines interfaces for user-function and
GraphQL functionality, and some convenience functions that invoke functions
or GraphQL through those interfaces.
The interfaces are implemented in the new 'functions' package.
The configuration types are also moved to the 'functions' package.
The 'rest' package imports 'functions' and takes care of compiling the config
and passing the compiled interface objects to the DB.