Support ad-hoc unit testing against existing Couchbase Server bucket
When running unit tests at dev time, there are scenarios where it's more efficient to test against an already provisioned bucket, instead of depending on the TestBucketPool to provision/flush the bucket.  To simplify this type of testing, this adds SG_TEST_USE_EXISTING_BUCKET environment variable which will bypass the bucket readier code, and instead attempt to establish a connection to the named bucket.