CBG-2333 move principal sequence handling to outside
Since currently there can only be a single collection, the sequence
allocator is shared between singleCollection.

An idea I had was to have two sequence allocators (regardless if there
are multiple collections or not):

(1) the principal sequence allocator
(2) the collection sequence allocator

This approach has some consequences - lastSequenceNumber from rest
API/stats will be different. I'd assumed we'd ignore the principal
sequences in reporting or report separately. (Current calls
`LastSequence`).

To use an actual separate sequence handler and have the tests pass, I
need `DatabaseCollection.WaitForPendingChanges` to wait for principal
sequences and non principal sequences. If I want to wait on a seqno, I'd
have to have a changeCache, and the changeCache is backed by a
DatabaseCollection, which we don't have for MetadataStore. We could
refactor this to be backed by a DataStore, if we do need to keep track
of these sequences.