Moved fns/GraphQL objects to Database
I added a js.VMPool to Database. Since UserFunctions and GraphQL require the
VMPool to initialize themselves, those objects can't be part of the
DatabaseContextOptions, which is created earlier.
So I defined an interface for the functions/GraphQL configuration (which is
defined up in the functions package) and put that on the Options instead.