CBG-2554 Channel computation and invalidation for collections
Extends the Principal and User interface to add collection-scoped API.

Adds Collections property to Authenticator config to allow callers to specify the set of collections that should be evaluated when channels are rebuilt.

Defines a new PrincipalCollectionAccess interface as a common interface for interaction with a collection's CollectionAccess that can be used for the default collection (stored on the principal root) or a named collection (stored in collection_access).