make TestCollectionsPutDocInKeyspace pass
The crux of the problem is that for base.Collection types
GetCollectionID would always return 0. Changed NamedDataStore to return
an optional error (implemented for gocb only), so we can report if we
can't get the collectionID from kv.

The problem with deferring GetCollectionID to happen later (in a synced
operation, so it would only happen) is that we have to handle the error
in a variety of unpleasant places, which is why I changed this in 0838041a2d4dd957f9bbfdbf7ad3440906ee5e56

This turns NamedDataStore into a less concrete type than it used to be.
There might be some type of caching to do here, TBD.