Collection support for NewUser in tests
NewUser is used by tests to instantiate a user with a set of admin-granted channels.  To support testing with and without named collections, NewUser has been updated to grant the specified channels to all collections defined for the authenticator.

There aren't any non-test use cases that were using NewUser with non-empty channels.  Converted these use cases to call NewUserNoChannels for the time being.  Should refactor NewUser to be test only, but it's preferable to handle that as a standalone change, given the number of tests impacted.