Turn off TCP 'NODELAY' option for WebSockets
Go's 'net' package sneakily(?) turns on the TCP 'NODELAY' option by default,
because it improves throughput in sockets that issue large writes.

However, NODELAY isn't so good with our replicator protocol because there are
a lot of small messages, and it causes each one to be sent as a separate packet.
This is especially noticeable with the replies to BLIP "rev" messages, which
are numerous and tiny.

This commit calls TCPConn.SetNoDelay(false) on sockets used for WebSockets.
It's a bit awkward to do because the TCP sockets are opened for us by the
'net/http' package, but fortunately there are hooks to get access to the
underlying TCPConn.