CBG-2343 Add database registry and associated API (#6101)
* Add database registry and associated API

Adds a new database registry document to the bucket that lists the set of databases defined for the bucket (across all config groups), and their associated collections.  The registry is used by persistent config polling to find all configs for a given bucket.

The registry is also used for cross-database config validation.  Currently the only validation is to ensure collections aren’t assigned to multiple databases (in separate config groups).

To synchronize access to the registry and ensure values are kept in step with the associated config files, changes were required to how the bootstrap manager interacts with config files.

The low-level config file operations in BootstrapConnection were made more generic so that they could be used for modifying both config and registry files.
A new API (ConfigManager), implemented by bootstrapContext in REST, should be used for all mutations to config files.  It includes handling for failure cases (including but not limited to node failure between registry and database config updates).

* License updates

* lint fixes

* More lint

* Test fixes

* Feedback tweak

* Additional test fixes

* working

* Fix unsuspendDatabase

UnsuspendDatabase was fetching and unmarshalling the database config on top of an already populated database config, which was the only reason credentials were being persisted across an unsuspend.

* Add migration for legacy configurations

Checks for legacy config storage on server context initialization - if found, migrates to 3.1 format.

* Error shadowing fix

* Simplify/clean-up error handling in GetDatabaseConfigs

* PR review fixes