CBG-2735 Compute metadataID when database is created, persist in registry
Adds KeyExists to BootstrapConnection interface to support checking existence of _sync:seq when determining whether to use the legacy metadata location.

References to _sync:syncInfo anticipate changes being made for CBG-2662 - decided to handle that work as a separate PR.

Requires the ability to run TestBucketPool using the default scope to test the fix.  Setting this via a new TestBucketPoolOptions struct, and it seemed to make sense to move MemWatermarkThresholdMB to the same struct.