Switch to forked graphql-go with CVE fix applied
Our dependency graphql-go v0.8.0 has a bug in which a malformed schema string
can cause a stack overflow in the parser, causing a Go panic. This is considered
a DoS attack vector, assigned CVE-2022-37315.

Seven months later, the fix for this bug has still not been merged, so we need
to fork graphql-go and apply the fix ourselves.

- Forked graphql-go repo to couchbasedeps/graphql-go.
- In the forked repo, cherry-picked fix of CVE-2022-37315, from
  https://github.com/graphql-go/graphql/pull/642 . Tagged this v0.8.1.
- Updated SG's go.mod file to override original go-graphql with our fork.
- Added a unit test in db/functions that tests the fix. I verified that,
  without the fix applied, this test panics; with it, it just returns an
  expected syntax error.