CBG-2662 Force resync when collection moves between databases
If a collection moves between databases, resync must be run with regenerateSequences=true in order to preserve sequence consistency.

When a collection is associated with a database, a syncInfo document (_sync:syncInfo) is written to the collection that stores the metadataID for the collection.

On database start, checks syncInfo for all collections.  On mismatch, starts the database in offline mode.

When in this state, a 'require_resync' property is included in the root db response and the root /_status endpoint with a slice of collections that require resync.

When resync w/ regenerate sequences is successfully run for a collection, syncInfo is updated for the collection.