Changes for docs
* Remove 'Admin only endpoints'
It is correct to be able to have multiple tags in OpenApi spec.
But it messes up our ability to categorise them with our *current*
toolset. Agreed with Ben Brooks and Isaac Lambat to remove for now,
but can revisit if you find you need the expressivity of multiple
tags.

* Fill in some details