Skip bucket Get when a new doc is pushed via BLIP
When a client pushes a doc and `proposeChanges` sees that the docID doesn't
exist in the bucket, the `rev` handler will write the doc to the bucket as new
without trying to get an existing copy first.

This will usually succeed, saving time. If it fails, it's handled like any
other CAS mismatch, by retrying and getting the existing doc.

- Added a field BlipSyncContext.pendingInsertions, a set of docIDs.
- RevDiff returns a new status `ProposedRev_OK_IsNew`.
- When the `proposeChanges` handler gets ProposedRev_OK_IsNew, it treats it
  like ProposedRev_OK and adds the docID to pendingInsertions.
- The `rev` handler then checks if the docID is in pendingInsertions; if so,
  it passes an empty BucketDoc with cas=0 to PutExistingRev as the
  `existingDoc`.
- WriteUpdateWithXattr now allows a `previous` doc with cas=0, interpreting it
  as the document not previously existing.