CBG-3031: [3.1.1] Allow one-shot replications to wait for DCP to catch up on changes feed (#6281)
* CBG-2853 Allow one-shot replications to wait for DCP to catch up on changes feed (#6243)

* CBG-2853 Add requestPlus option for changes feeds

Adds requestPlus option for changes feeds.  When set, changes feeds will loop until the cached sequence (via DCP) is greater than the database sequence at the time the changes request was issued.

requestPlus can be enabled for non-continuous changes requests in one of three ways:
- by setting request_plus=true on a REST API changes call
- by setting the requestPlus property to "true" on a subChanges message
- by setting "changes_request_plus":true in the database config (default=false)

The request setting is given priority - if not set on a request, the value will fall back to the database config value.

Required minor refactoring of how options.Wait was used in changes.go, to support use of requestPlus and longpoll together.  No functional changes to longpoll if requestPlus is not set.

* Update docs for request_plus changes parameter.

* lint fixes

* Fix race in TestOneShotGrant tests (#6247)

---------

Co-authored-by: Adam Fraser <adam.fraser@couchbase.com>