CBG-3291 Define macro expansion in MutateInOpts
Uptake sg-bucket changes to pass macro expansions for a given operation to the XattrStore implementation (*Collection), instead of hardcoding in the low-level subdoc operations.

The macro expansion options are set in three places for non-test code:
- updateAndReturnDoc, handles standard writes/imports via WriteUpdateWithXattr
- migrateMetadata, performs migration via WriteWithXattr
- resyncDocument, uses WriteUpdateWithXattr (different retry than updateAndReturnDoc)

Prerequisite for the expanded use of macro expansion for HLV properties.

Also includes some cleanup of the XattrStore interface to remove internal-only operations.