CBG-3498 make test pass without named collections
Revert accidental changes to channels assertion. The response to PUT
/db/_session will only show the channels of the default collection. We don't
show the channels of any non default collections, so if we are using named
collections, this value is nil.

This preserves 3.0 behavior. The channels property is also misleading because
it doesn't show the inherited channels from roles, but this is preserving 3.0
behavior and not introducing any slowdown by calculating inherited
channels.